feat(tests): Fix tests teardown and setup methods
