+++ s3-agent/tests/requirements.txt
@@ -1,0 +1,1 @@
+pytest >= 7.1.2, <= 8.0.0
