--- tests/requirements.txt
@@ -1,1 +1,0 @@
-pytest==3.4.2
